update profil terbaru

+++ application/controllers/Barang.php
@@ -56,7 +56,7 @@ class Barang extends CI_Controller {
 			echo "window.location='".site_url('barang')."';</script>";
 		}
 	}
-
+	
 	public function proses(){
 		$config['upload_path']          = './assets/gambar/barang';
 		$config['allowed_types']        = 'jpeg|jpg|png';
@@ -65,30 +65,50 @@ class Barang extends CI_Controller {
 		
 		$post = $this->input->post(null, TRUE);
 		if(isset($_POST['tambah'])){
-			if($this->upload->do_upload('gambar')){
-				$post['gambar'] = $this->upload->data('file_name');
-				$this->barang_model->add($post);
-			}else{
-				echo "upload gagal";
-			}
-		}else if(isset($_POST['ubah'])){
-			if($this->upload->do_upload('gambar')){
-				$item = $this->barang_model->get($post['id_brg'])->row();
-				if($item->gambar != null){
-					$target_file = './assets/gambar/barang/'.$item->gambar;
-					unlink($target_file);
+			if($_FILES['gambar']['name']!=null){
+				if($this->upload->do_upload('gambar')){
+					$post['gambar'] = $this->upload->data('file_name');
+					$this->barang_model->add($post);
+					echo "<script>alert('Data Berhasil Disimpan');</script>";
+					echo "<script>window.location='".site_url('barang')."';</script>";
+				}else{
+					echo "<script>alert('Data Gagal Disimpan');</script>";
+					echo "<script>window.location='".site_url('barang')."';</script>";
 				}
-
-				$post['gambar'] = $this->upload->data('file_name');
-				$this->barang_model->edit($post);
-			}else{
-				echo "upload gagal";
+			} else {
+				$post['foto'] = null;
+				$this->barang_model->add($post);
+				echo "<script>
+				alert('Data berhasil disimpan');
+				window.location='".site_url('barang')."';
+				</script>";
 			}
-		} 
-		if($this->db->affected_rows()>0){
-			echo "<script>alert('Data Berhasil Disimpan');</script>";
-		}
-		echo "<script>window.location='".site_url('barang')."';</script>";
+		
+		}else if(isset($_POST['ubah'])){
+			if($_FILES['gambar']['name']!=null){
+				if($this->upload->do_upload('gambar')){
+					$item = $this->barang_model->get($post['id_brg'])->row();
+					if($item->gambar != null){
+						$target_file = './assets/gambar/barang/'.$item->gambar;
+						unlink($target_file);
+					}
 
+					$post['gambar'] = $this->upload->data('file_name');
+					$this->barang_model->edit($post);
+					echo "<script>alert('Data Berhasil Disimpan');</script>";
+					echo "<script>window.location='".site_url('barang')."';</script>";
+				}else{
+					echo "<script>alert('Data Gagal Disimpan');</script>";
+					echo "<script>window.location='".site_url('barang')."';</script>";
+				}
+			} else {
+				$post['gambar'] = null;
+				$this->barang_model->edit($post);
+				echo "<script>
+				alert('Data berhasil disimpan');
+				window.location='".site_url('barang')."';
+				</script>";
+			}
+		}
 	}
 }
